agents/agent-oo/pyproject.toml
[project]
name = "agent-waa-3"
version = "0.0.1"
description = "Agent based on WAA"
authors = [{name = "Lukas Kellerstein", email = "lkellerstein@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.4.0",
    "addict>=2.4.0",
    "anthropic>=0.47.2",
    "anytree>=2.12.1",
    "backoff>=2.2.1",
    "beautifulsoup4>=4.13.3",
    "borb>=2.1.25",
    "chardet>=5.2.0",
    "cssselect>=1.2.0",
    "dashscope>=1.22.1",
    "easyocr>=1.7.2",
    "einops>=0.8.1",
    "fabric>=3.2.2",
    "fastdtw>=0.3.4",
    "flask~=3.0.0",
    "formulas>=1.2.8",
    "func-timeout>=4.3.5",
    "gdown>=5.2.0",
    "google-generativeai>=0.8.4",
    "groq>=0.18.0",
    "gymnasium~=0.28.1",
    "imagehash>=4.3.2",
    "librosa>=0.10.2.post1",
    "lnkparse3>=1.5.0",
    "lxml>=5.3.1",
    "matplotlib~=3.7.4",
    "mutagen>=1.47.0",
    "numpy>=1.24.5",
    "odfpy>=1.4.1",
    "openai>=1.64.0",
    "opencv-python~=4.8.1.78",
    "openpyxl>=3.1.5",
    "pandas~=2.2.3",
    "pdfplumber>=0.11.5",
    "pillow~=10.3.0",
    "playwright>=1.50.0",
    "psutil~=5.9.6",
    "py7zr>=0.22.0",
    "pyacoustid>=1.3.0",
    "pyautogui~=0.9.54",
    "pycocotools>=2.0.8",
    "pydrive>=1.3.1",
    "pygetwindow>=0.0.9",
    "pymupdf>=1.25.3",
    "pynput~=1.7.6",
    "pypdf>=5.3.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "python-pptx>=1.0.2",
    "pyyaml>=6.0.2",
    "rapidfuzz>=3.12.1",
    "requests~=2.32.0",
    "requests-toolbelt~=1.0.0",
    "scikit-image>=0.25.2",
    "supervision>=0.24.0",
    "tenacity>=9.0.0",
    "tiktoken>=0.9.0",
    "timm>=1.0.15",
    "torch~=2.2.0",
    "torchvision>=0.17.2",
    "tqdm~=4.65.0",
    "transformers~=4.46.1",
    "ultralytics>=8.3.80",
    "wandb>=0.19.7",
    "wrapt-timeout-decorator>=1.5.1",
    "xmltodict>=0.14.2",
    "yapf>=0.43.0",
]
